Move assign_if_not_finite to utils
