+++ stingray/modeling/posterior.py
@@ -15,6 +15,7 @@ from astropy.modeling.fitting import _fitter_to_model_params
 from astropy.modeling import models
 
 from stingray import Lightcurve, Powerspectrum
+from stingray.utils import assign_if_not_finite
 
 
 # TODO: Add checks and balances to code
@@ -27,42 +28,6 @@ __all__ = ["set_logprior", "Posterior", "PSDPosterior", "LogLikelihood", "Poisso
            "PriorUndefinedError", "LikelihoodUndefinedError"]
 
 logmin = -10000000000000000.0
-
-
-def assign_if_not_finite(value, default):
-    """Check if a value is finite. Otherwise, return the default.
-
-    Parameters
-    ----------
-    value : float, int or `np.array`
-        The input value
-    default : float
-        The default value
-
-    Returns
-    -------
-    output : same as ``value``
-        The result
-
-    Examples
-    --------
-    >>> assign_if_not_finite(1, 3.2)
-    1
-    >>> assign_if_not_finite(np.inf, 3.2)
-    3.2
-    >>> input_arr = np.array([np.nan, 1, np.inf, 2])
-    >>> np.allclose(assign_if_not_finite(input_arr, 3.2), [3.2, 1, 3.2, 2])
-    True
-
-    """
-    if isinstance(value, Iterable):
-        values = [assign_if_not_finite(val, default) for val in value]
-        values = np.array(values)
-        return values
-
-    if not np.isfinite(value):
-        return default
-    return value
 
 
 class PriorUndefinedError(Exception):
